helpers.el: ditch insert-evaluation-steps in favor of region-to-evaluation-steps
Exercise 48 requires a more complex form that calls for multiline
editing.

I considered writing a function that pops up a temp buffer (with,
e.g., `read-string-from-buffer`) but felt that was too much
trouble (mainly because I only stumbled upon the latter function well
into writing this hah). So I figured I was already in the mode that I
wanted to be in, and could just leverage that to hackily turn a series
of forms into steps :shrug: