helpers.el: (kinda) fix region-to-evaluation-steps
and name it paragraph-to-evaluation-steps

This would behave inconsistently, probably because the point shifts as
we add the == comments. This takes a simpler approach: it simply
assumes that we want to do this for a paragraph (i.e. up til an empty
line).

This is still kinda sad, though, because you wind up adding an ==
comment after the last sexp, which defeats the whole original idea of
having this replace the original helper, which was interactive.